Extract validations specific class
